Fixed PermissionGroup not being able to hold more than one group of the same priority
Fixed PermissionGroup not being able to hold more than one group of the same priority.

Replaced the referencedGroups TreeSet with a sorted list, which should be more suited to this use case. Updated .addPermissionGroup to accommodate this.

Made priorityComparatorHighestFirst consider the permission group's name as well, so groups are in alphabetical order after priority order.

Removed priorityComparatorLowestFirst, as it's not needed and its functionality would now be ambiguous. (Should it also use reverse alphabetical order?)