Added support for multi-line permission args.
Added support for multi-line permission args.

Permission args are considered to continue to the next line where the next line is at least 4 spaces indented compared to the permission it's an arg of.

Implemented my own line reader to handle treating a permission with a multi-line arg as one "line".

Added part to PermissionSet.applyPermissionToPathString to, where a permission arg is multi-line, start it on the following line and increment the arg appropriately.

Added PermissionSet.setWhileDeIndenting and PermissionGroup.addPermissionWhileDeIndenting to facilitate multi-line permission args being de-indented appropriately, and utilised them.