Added and implemented ability for PermissionsRegistry to provide users, groups, and user/group permissions
Added and implemented PermissionsRegistry.getGroupNames(), .getUsers(), .getUserPermissions(...), .getGroupPermissions(...), and internal .getPermissions(...) methods.

Removed leftover comment in PermissionsRegistry.

Added and implemented PermissionGroup.getPermissionsAsStrings(...)

Added and implemented PermissionSet.getPermissionsAsStrings. Rewrote part of PermissionSet to allow for the provision of contained permissions as strings, and to allow specification of whether or not to include permission args in those strings.

Added tests to cover new functionality in PermissionsRegistry.