Added the ability for PermissionsRegistry to have default permissions for users.
Added the ability for PermissionsRegistry to have a default permission group for users. Default permissions are checked in the same way as the groups a user might have, after all other groups.

Added and implemented PermissionGroup constructors to accept a default permission group.

Added a default empty permission group as a static final variable on PermissionGroup, which doesn't have any permissions or groups. This is required to prevent an infinite loop as PermissionGroups now require a default permission group on construction.

Made PermissionGroup.getMostRelevantPermission(...) defer to the default permission group failing everything else.

Added and implemented PermissionGroup.isEmpty() and .clear()

Added and implemented PermissionSet.clear()

Gave PermisisonsRegistry the ability to save and load the default permissions.

Added and implemented in PermissionsRegistry:
.getDefaultPermissions()
.getDefaultGroups()
.assignDefaultPermission(...),
.revokeDefaultPermission(...),
.assignDefaultGroup(...),
.revokeDefaultGroup(...),
.hasDefaultPermission(...),
.isDefaultGroup(...),
.getDefaultPermissionArg(...)

Made group equivalents for the above methods defer to them when the group ID is given as "*".

Also, made PermissionsRegistry.createGroup(...) and .createGroupFromSaveString(...) not public - they expose PermissionsGroup to the public interface.