Removed support for conditional permissions.
Stripped support for conditional permissions from codebase.

This feature was adding too much complexity compared to the utility and clouding the definition of certain concept of the system, and continuing the implementation properly would have resulted in classes being coupled when they shouldn't be. Comparable functionality can be implemented is a simpler way, I'm thinking as a decorator that allows ensuring that users do or do not have permissions based on external criteria.

Removed tests relating to this as well.

Also:
 - Fixed a PermissionGroup test.
 - Added and utilised a constructor for PermissionSet.PermissionWithPath that accepts an appropriate TreeEntry, just as a convenience.