Added methods to PermissionsRegistry dealing with priority, making multiple adjustments at once, and getting permissions as strings with args.
Created the static inner class of PermissionGroup.Priority. This would normally be a named typed tuple. It contains both possible representations of a permission group's priority, and which one it is intended to be represented as.

In PermissionGroup, added .getPriorityAsObject() for getting the groups' priority as of the the aforementioned wrappers.

In PermissionsRegistry, added methods allowing for accessing group priority, string representations of permissions with args directly, and adding multiple permissions/groups in one operation. Added:
 - .getGroupPriority(...)
 - .getGroupPriorityAsLong(...)
 - .getGroupPriorityAsObject(...)
 - .getUserPermissionsWithArgs(...)
 - .getGroupPermissionsWithArgs(...)
 - .getDefaultPermissionsWithArgs()
 - .assignUserPermissions(...)
 - .assignGroupPermissions(...)
 - .assignDefaultPermissions(...)
 - .assignPermissions(...)
 - .assignGroupsToUser(...)
 - .assignGroupsToGroup(...)
 - .assignDefaultGroups(...)
 - .assignGroupsTo(...)

Propagated these changes to PermissionsRegistryDecorator and its subclasses.

Added new assignment methods to PermissionsRegistryWithEvents, although they're currently not implemented.

Also:
 - Changed some more instances of "group IDs" to "group names"
 - Fixed small bug in PermissionGroup's .toSaveString() method.