Separated PermissionsRegistry into interfaces and implementation
Separated PermissionsRegistry into interfaces and implementation. Now PermissionsRegistry, GroupBasedPermissionsRegistry, and GroupMapPermissionsRegistry.

Code has been refactored to account for this, and tests updated to reflect this. Some tests are now not fully implemented and need to be updated.

PermissionsRegistryDecorator current decorates instances of GroupBasedPermissionsRegistry, rather than than instances of the interface PermissionsRegistry. This needs to be changed.